requirements.txt
hikari==2.0.0.dev106
PyYAML~=6.0
hikari-lightbulb
hikari-miru
XlsxWriter~=3.0.2
pytz~=2021.3
